#  proc/variables: prefetch of target process memory 