--- a/_fixtures/testvariables3.go
+++ b/_fixtures/testvariables3.go
@@ -48,6 +48,11 @@ type dstruct struct {
 
 type maptype map[string]interface{}
 
+type benchstruct struct {
+	a [64]byte
+	b [64]byte
+}
+
 func main() {
 	i1 := 1
 	i2 := 2
@@ -138,6 +143,12 @@ func main() {
 	var iface2fn2 interface{} = afunc2
 	var mapinf maptype = map[string]interface{}{}
 	mapinf["inf"] = mapinf
+	var bencharr [64]benchstruct
+	var benchparr [64]*benchstruct
+
+	for i := range benchparr {
+		benchparr[i] = &benchstruct{}
+	}
 
 	var amb1 = 1
 	runtime.Breakpoint()
@@ -145,5 +156,5 @@ func main() {
 		fmt.Println(amb1)
 	}
 	runtime.Breakpoint()
-	fmt.Println(i1, i2, i3, p1, amb1, s1, s3, a1, p2, p3, s2, as1, str1, f1, fn1, fn2, nilslice, nilptr, ch1, chnil, m1, mnil, m2, m3, up1, i4, i5, i6, err1, err2, errnil, iface1, iface2, ifacenil, arr1, parr, cpx1, const1, iface3, iface4, recursive1, recursive1.x, iface5, iface2fn1, iface2fn2, mapinf)
+	fmt.Println(i1, i2, i3, p1, amb1, s1, s3, a1, p2, p3, s2, as1, str1, f1, fn1, fn2, nilslice, nilptr, ch1, chnil, m1, mnil, m2, m3, up1, i4, i5, i6, err1, err2, errnil, iface1, iface2, ifacenil, arr1, parr, cpx1, const1, iface3, iface4, recursive1, recursive1.x, iface5, iface2fn1, iface2fn2, bencharr, benchparr, mapinf)
 }