@derekparker 2016-01-05
This is great, I'm very excited about this patch. Is there any reason this is on top of #315? That PR has a legitimate failure on OSX, which is causing CI to fail on this PR.

I will review the code tomorrow morning.