@aarzilli 2016-01-05
> This is great, I'm very excited about this patch. Is there any reason this is on top of #315? 

No, it's a mistake. I also didn't notice the issue with #315