@aarzilli 2016-01-05
Decoupled from #315